bugfix: file size was counted in int, changed to long

+++ src/main/java/com/github/n_i_e/dirtreedb/PathEntry.java
@@ -210,7 +210,7 @@ public class PathEntry {
 
 	public void setCsum(InputStream inf) throws IOException {
 		try {
-			int contentsize = 0;
+			long contentsize = 0;
 			MessageDigest md5digest = MessageDigest.getInstance("MD5");
 
 			byte[] buff = new byte[4096];
